+++ packages/jsonapi/src/Requests/CreationRequest.php
@@ -37,13 +37,13 @@ class CreationRequest
         $requestBody = $this->requestTransformer->getCreationRequestBody($typeName, $expectedProperties);
         $urlParams = $this->requestTransformer->getUrlParameters();
 
-        $beforeCreationEvent = new BeforeCreationEvent($type);
+        $beforeCreationEvent = new BeforeCreationEvent($type, $requestBody);
         $this->eventDispatcher->dispatch($beforeCreationEvent);
 
         $modifiedEntity = $type->createEntity($requestBody);
         $entity = $modifiedEntity->getEntity();
 
-        $afterCreationEvent = new AfterCreationEvent($type, $entity);
+        $afterCreationEvent = new AfterCreationEvent($type, $entity, $requestBody);
         $this->eventDispatcher->dispatch($afterCreationEvent);
 
         $requestDeviations = array_merge(