+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 
 # Changelog
 
+## Unreleased
+
+* feature: provide request body in creation and update events
+
 ## 0.24.29 - 2023-12-12
 
 * feature: rework "instance of" implementation to be properly usable (though support for execution in `ConditionEvaluator` is still missing)